# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Mané API",
    "version": "1.2.1"
  },
  "servers": [{ "url": "/" }],
  "paths": {
...
    "/v1/units": {
      "get": {
        "summary": "List units",
        "description": "Público. Use `active=true` para listar apenas unidades ativas.",
        "parameters": [
          { "in": "query", "name": "page", "schema": { "type": "integer", "minimum": 1 }, "required": false },
          { "in": "query", "name": "pageSize", "schema": { "type": "integer", "minimum": 1, "maximum": 100 }, "required": false },
          { "in": "query", "name": "search", "schema": { "type": "string" }, "required": false },
          { "in": "query", "name": "active", "schema": { "type": "boolean" }, "required": false }
        ],
        "responses": {
          "200": { "description": "ok", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UnitListResponse" } } } }
        }
      },
      "post": {
        "summary": "Create unit",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UnitCreate" } } }
        },
        "responses": {
          "201": { "description": "created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Unit" } } } },
          "400": { "description": "bad request" },
          "401": { "description": "unauthorized" },
          "403": { "description": "forbidden" },
          "409": { "description": "slug already in use" }
        }
      }
    },
...
    "/v1/units/{id}": {
      "get": {
        "summary": "Get unit by id",
        "description": "Público.",
        "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "ok", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Unit" } } } },
          "404": { "description": "not found" }
        }
      },
      "put": {
        "summary": "Update unit",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UnitUpdate" } } }
        },
        "responses": {
          "200": { "description": "ok", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Unit" } } } },
          "400": { "description": "bad request" },
          "401": { "description": "unauthorized" },
          "403": { "description": "forbidden" },
          "404": { "description": "not found" },
          "409": { "description": "slug already in use" }
        }
      },
      "delete": {
        "summary": "Delete unit",
        "description": "Hard delete. Bloqueado caso existam reservas associadas.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "204": { "description": "no content" },
          "401": { "description": "unauthorized" },
          "403": { "description": "forbidden" },
          "409": { "description": "unit has linked reservations" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Cole o token gerado por /auth/login. No Swagger UI, use o JWT completo (sem a palavra 'Bearer')."
      }
    },
    "schemas": {
...
      "Unit": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "slug": { "type": "string" },
          "isActive": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "UnitCreate": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": { "type": "string", "example": "Águas Claras" },
          "slug": { "type": "string", "example": "aguas-claras" },
          "isActive": { "type": "boolean", "example": true }
        }
      },
      "UnitUpdate": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "slug": { "type": "string" },
          "isActive": { "type": "boolean" }
        }
      },
      "UnitListResponse": {
        "type": "object",
        "properties": {
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/Unit" } },
          "total": { "type": "integer" },
          "page": { "type": "integer" },
          "pageSize": { "type": "integer" },
          "totalPages": { "type": "integer" }
        }
      }
    }
  }
}